+++ Plugin/CleverTapUnity/CleverTapUnity-Scripts/CleverTapBinding.cs
@@ -11,7 +11,7 @@ using CleverTap.Utilities;
 namespace CleverTap {
   public class CleverTapBinding : MonoBehaviour {
       
-    public const string Version = "2.3.1";
+    public const string Version = "2.4.0";
 
 #if UNITY_IOS
     void Start() {
@@ -233,6 +233,15 @@ namespace CleverTap {
 
     [System.Runtime.InteropServices.DllImport("__Internal")]
     private static extern bool CleverTap_getFeatureFlag(string key, bool defaultValue);
+
+    [System.Runtime.InteropServices.DllImport("__Internal")]
+    private static extern void CleverTap_promptForPushPermission(bool showFallbackSettings);
+
+    [System.Runtime.InteropServices.DllImport("__Internal")]
+    private static extern void CleverTap_promptPushPrimer(string json);
+
+    [System.Runtime.InteropServices.DllImport("__Internal")]
+    private static extern void CleverTap_isPushPermissionGranted();
 
     public static void LaunchWithCredentials(string accountID, string token) {
         CleverTap_launchWithCredentials(accountID, token);
@@ -610,6 +619,19 @@ namespace CleverTap {
         return CleverTap_getFeatureFlag(key, defaultValue);
     }
 
+    public static void PromptPushPrimer(Dictionary<string, object> json) {
+        var jsonString = Json.Serialize(json);
+        CleverTap_promptPushPrimer(jsonString);
+    }
+
+
+    public static void PromptForPushPermission(bool showFallbackSettings) {
+        CleverTap_promptForPushPermission(showFallbackSettings);
+    }
+
+    public static void IsPushPermissionGranted() {
+        CleverTap_isPushPermissionGranted();
+    }
 
 #elif UNITY_ANDROID
     private static AndroidJavaObject unityActivity;
@@ -864,6 +886,19 @@ namespace CleverTap {
     public static void ProfileRemoveMultiValueForKey(string key, string val) {
         CleverTap.Call("profileRemoveMultiValueForKey", key, val);
     }
+
+    public static void SuspendInAppNotifications() {
+        CleverTap.Call("suspendInAppNotifications");
+    }
+
+    public static void DiscardInAppNotifications() {
+        CleverTap.Call("discardInAppNotifications");
+    }
+
+    public static void ResumeInAppNotifications() {
+        CleverTap.Call("resumeInAppNotifications");
+    }
+
     public static void RecordScreenView(string screenName) {
         CleverTap.Call("recordScreenView", screenName);
     }
@@ -976,6 +1011,17 @@ namespace CleverTap {
         return CleverTap.Call<int>("getInboxMessageUnreadCount");
     }
 
+    public static void PromptPushPrimer(Dictionary<string, object> details){
+         CleverTap.Call("promptPushPrimer", Json.Serialize(details));
+    }
+
+    public static void PromptForPushPermission(bool showFallbackSettings){
+         CleverTap.Call("promptForPushPermission", showFallbackSettings);
+    }
+
+    public static bool IsPushPermissionGranted(){
+        return CleverTap.Call<bool>("isPushPermissionGranted");
+    }
 #else
 
    // Empty implementations of the API, in case the application is being compiled for a platform other than iOS or Android.
@@ -1164,6 +1210,15 @@ namespace CleverTap {
 
     public static int GetInboxMessageUnreadCount(){
         return -1;
+    }
+
+    public static void PromptPushPrimer(string json){
+    }
+
+    public static void PromptForPushPermission(bool showFallbackSettings){
+    }
+
+    public static void IsPushPermissionGranted(){
     }
 #endif
     }