+++ example/CleverTapUnity.cs
@@ -57,6 +57,45 @@ public class CleverTapUnity : MonoBehaviour
         CleverTapBinding.RecordEvent("Test Unity Event");
         //Invoke("LaunchInbox",30.0f);
 
+        //Push primer APIs usage
+
+        CleverTapBinding.IsPushPermissionGranted();
+
+        //Half-Interstial Local InApp
+        Dictionary<string, object> item = new Dictionary<string, object>();
+        item.Add("inAppType", "half-interstitial");
+        item.Add("titleText", "Get Notified");
+        item.Add("messageText", "Please enable notifications on your device to use Push Notifications.");
+        item.Add("followDeviceOrientation", true);
+        item.Add("positiveBtnText", "Allow");
+        item.Add("negativeBtnText", "Cancel");
+        item.Add("backgroundColor", "#FFFFFF");
+        item.Add("btnBorderColor", "#0000FF");
+        item.Add("titleTextColor", "#0000FF");
+        item.Add("messageTextColor", "#000000");
+        item.Add("btnTextColor", "#FFFFFF");
+        item.Add("btnBackgroundColor", "#0000FF");
+        item.Add("imageUrl", "https://icons.iconarchive.com/icons/treetog/junior/64/camera-icon.png");
+        item.Add("btnBorderRadius", "2");
+        item.Add("fallbackToSettings", true);
+        CleverTapBinding.PromptPushPrimer(item);
+
+        //Alert Local InApp
+        Dictionary<string, object> item = new Dictionary<string, object>();
+        item.Add("inAppType", "half-interstitial");
+        item.Add("titleText", "Get Notified");
+        item.Add("messageText", "Please enable notifications on your device to use Push Notifications.");
+        item.Add("followDeviceOrientation", true);
+        item.Add("fallbackToSettings", true);
+        CleverTapBinding.PromptPushPrimer(item);
+
+        /*Prompt to show hard notification permission dialog
+          true - fallbacks to app's notification settings if permission is denied
+          false - does not fallback to app's notification settings if permission is denied
+        */
+        CleverTapBinding.PromptForPushPermission(false);
+
+
     }
 
     /* --------------------------------------------------------------------------------
@@ -419,6 +458,17 @@ public class CleverTapUnity : MonoBehaviour
         Debug.Log("unity received inbox initialized");
         // string[] arr = {"1608798445_1676289520","1548315289_1676289520","1608798445_1676289268", "1548315289_1676289268"};
         // CleverTapBinding.DeleteInboxMessagesForIDs(arr);
+    }
+
+    // returns the status of push permission response after it's granted/denied
+    void CleverTapOnPushPermissionResponseCallback(string message) {
+        Debug.Log("unity received push permission response: " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
+    }
+
+    // returns callback to indicate whether notification permission is granted or not
+    void CleverTapPushNotificationPermissionStatus(string message)
+    {
+        Debug.Log("CleverTap application isPushPermissionGranted " + (!String.IsNullOrEmpty(message) ? message : "NULL"));
     }
 
     void CleverTapInboxMessagesDidUpdateCallback()