+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Change Log
 ==========
+Version 2.4.0 *(XX April, 2023)*
+-------------------------------------------
+- Adds below new public APIs to support [CleverTap Android SDK v4.7.5](https://github.com/CleverTap/clevertap-android-sdk/releases/tag/corev4.7.5_rmv1.0.3)
+    - `CleverTapBinding.IsPushPermissionGranted()`, `CleverTapBinding.PromptPushPrimer(object)`, `CleverTapBinding.PromptForPushPermission(boolean)`
+- Adds push permission callback method `CleverTapOnPushPermissionResponseCallback` which returns true/false after user allows/denies the notification permission.
+- Adds `CleverTapInAppNotificationShowCallback` to handle InApp notification shown - Only for Android.
 
 Version 2.3.1 *(13 April, 2023)*
 -------------------------------------------